This is a combination of 2 commits.
This is a combination of 2 commits.

This is a combination of 2 commits.

This is a combination of 2 commits.

This is a combination of 2 commits.

This is a combination of 2 commits.

Initial push for problem 88 and building template architecture

Error build state with C++

Moved bash scripts and re-org folder

Fixed CMakeLists.txt to find interface library

Updated solution to match template

Updated template functions of Solutions class

Added linting as a task